requirements - Copy.txt
streamlit
pymupdf
sentence-transformers
faiss-cpu
transformers
accelerate
torch
